shuffle exports between store and types
