+++ frontend/src/Attachment/ReplacePart.svelte
@@ -5,13 +5,13 @@
     ModalHeader,
     ModalBody,
   } from "@sveltestrap/sveltestrap";
-  import { Type, maxDate } from "../lib/types";
   import { AttEvent, Attachment } from "../lib/attachment";
-  import { types, user } from "../lib/store";
+  import { user, maxDate } from "../lib/store";
+  import { types, Type } from "../lib/types";
+  import { parts, Part } from "../lib/part";
   import ModalFooter from "../Widgets/ModalFooter.svelte";
   import NewForm from "../Part/PartForm.svelte";
   import Dispose from "../Widgets/Dispose.svelte";
-  import { parts, Part } from "../lib/part";
 
   let part: Part, oldpart: Part, newpart: Part;
   let type: Type;
