+++ frontend/src/ServicePlan/PlanHook.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type { ServicePlan } from "../lib/serviceplan";
-  import { types } from "../lib/store";
+  import { types } from "../lib/types";
   import { link } from "svelte-spa-router";
   import { attachments, part_at_hook } from "../lib/attachment";
