+++ frontend/src/lib/types.ts
@@ -1,8 +1,8 @@
 import { Activity } from "./activity";
 import { Part } from "./part";
-import { type Map, by, filterValues } from "./mapable";
-import { types } from "./store";
-export const maxDate = new Date("2999-12-31");
+import { type Map, by, filterValues, mapObject } from "./mapable";
+import { myfetch } from "./store";
+import { writable, type Writable } from "svelte/store";
 
 export class Type {
   id: number;
@@ -54,3 +54,25 @@ export type ActType = {
   name: string;
   gear_type: number;
 };
+
+export let types: Map<Type>;
+
+export async function getTypes() {
+  let res = Promise.all([
+    myfetch("/api/types/part").then((types) =>
+      types.map((t: any) => new Type(t)).reduce(mapObject("id"), {}),
+    ), // data[0]
+    myfetch("/api/types/activity"), // data[1]
+  ])
+    .then((data: { 0: Map<Type>; 1: ActType[] }) => {
+      types = data[1].reduce((acc, a) => {
+        acc[a.gear_type].acts.push(a);
+        return acc;
+      }, data[0]);
+    })
+    .then(() => (category = writable(types[1])));
+  console.log(types);
+  return res;
+}
+
+export let category: Writable<Type>;
